pipeline_reviews/requirements.txt
nltk
pandas
psycopg2-binary
python-dotenv
requests